+++ core/cat/factory/embedder.py
@@ -39,6 +39,24 @@ class EmbedderOpenAIConfig(EmbedderSettings):
         }
 
 
+# https://python.langchain.com/en/latest/_modules/langchain/embeddings/openai.html#OpenAIEmbeddings
+class EmbedderAzureOpenAIConfig(EmbedderSettings):
+    openai_api_key: str
+    model: str
+    openai_api_base: str
+    api_type: str
+    api_version: str
+    deployment: str
+
+    _pyclass: PyObject = langchain.embeddings.OpenAIEmbeddings
+
+    class Config:
+        schema_extra = {
+            "name_human_readable": "Azure OpenAI Embedder",
+            "description": "Configuration for Azure OpenAI embeddings",
+        }
+
+
 class EmbedderCohereConfig(EmbedderSettings):
     cohere_api_key: str
     _pyclass: PyObject = langchain.embeddings.CohereEmbeddings
@@ -65,6 +83,7 @@ class EmbedderHuggingFaceHubConfig(EmbedderSettings):
 SUPPORTED_EMDEDDING_MODELS = [
     EmbedderFakeConfig,
     EmbedderOpenAIConfig,
+    EmbedderAzureOpenAIConfig,
     EmbedderCohereConfig,
     EmbedderHuggingFaceHubConfig,
 ]