+++ core/cat/looking_glass/cheshire_cat.py
@@ -1,14 +1,14 @@
 import time
 import traceback
 from typing import Union
-from datetime import timedelta
 
 import langchain
-from cat.utils import log, verbal_timedelta
+from cat.utils import log
 from cat.db.database import get_db_session, create_db_and_tables
 from cat.rabbit_hole import RabbitHole
 from starlette.datastructures import UploadFile
 from cat.mad_hatter.mad_hatter import MadHatter
+from cat.memory.working_memory import WorkingMemory
 from cat.memory.long_term_memory import LongTermMemory
 from langchain.docstore.document import Document
 from cat.looking_glass.agent_manager import AgentManager
@@ -38,7 +38,9 @@ class CheshireCat:
         self.load_plugins()
         self.load_natural_language()
         self.load_memory()
-        self.load_agent()
+
+        # Agent manager instance (for reasoning)
+        self.agent_manager = AgentManager(self)
 
         # Rabbit Hole Instance
         self.rabbit_hole = RabbitHole()
@@ -54,10 +56,6 @@ class CheshireCat:
         # LLM and embedder
         self.llm = self.mad_hatter.execute_hook("get_language_model")
         self.embedder = self.mad_hatter.execute_hook("get_language_embedder")
-
-        # Prompts
-        self.prefix_prompt = self.mad_hatter.execute_hook("get_main_prompt_prefix")
-        self.suffix_prompt = self.mad_hatter.execute_hook("get_main_prompt_suffix")
 
         # HyDE chain
         hypothesis_prompt = langchain.PromptTemplate(
@@ -95,59 +93,11 @@ class CheshireCat:
         # Memory
         vector_memory_config = {"embedder": self.embedder, "verbose": True}
         self.memory = LongTermMemory(vector_memory_config=vector_memory_config)
-
-    def reset_history(self):
-        # clear recent conversation
-        self.history = ""
+        self.working_memory = WorkingMemory()
 
     def load_plugins(self):
-        # recent conversation
-        # TODO: use working memory to manage convo history
-        self.reset_history()
-
         # Load plugin system
         self.mad_hatter = MadHatter(self)
-
-    def load_agent(self):
-        self.agent_manager = AgentManager(
-            llm=self.llm,
-            tools=self.mad_hatter.tools,
-            verbose=self.verbose,
-        )  # TODO: load agent from plugins? It's gonna be a MESS
-
-        self.agent_executor = self.agent_manager.get_agent_executor(
-            prefix_prompt=self.prefix_prompt,
-            suffix_prompt=self.suffix_prompt,
-            # ai_prefix="AI",
-            # human_prefix="Human",
-            input_variables=self.input_variables,
-            return_intermediate_steps=True,
-        )
-
-    # TODO: this should be a hook
-    def format_memories_for_prompt(self, memory_docs, return_format=str):
-        memory_texts = [m[0].page_content.replace("\n", ". ") for m in memory_docs]
-
-        # TODO: take away duplicates
-        memory_timestamps = []
-        for m in memory_docs:
-            timestamp = m[0].metadata["when"]
-            delta = timedelta(seconds=(time.time() - timestamp))
-            memory_timestamps.append(" (" + verbal_timedelta(delta) + ")")
-
-        memory_texts = [a + b for a, b in zip(memory_texts, memory_timestamps)]
-        # TODO: insert sources in document memories
-
-        if return_format == str:
-            memories_separator = "\n  - "
-            memory_content = memories_separator + memories_separator.join(memory_texts)
-        else:
-            memory_content = memory_texts
-
-        if self.verbose:
-            log(memory_content)
-
-        return memory_content
 
     def get_hyde_text_and_embedding(self, text):
         # HyDE text
@@ -170,19 +120,23 @@ class CheshireCat:
 
         # loop until there are no groups to summarize
         root_summary_flag = False
+        separator = "\n --> "
         while not root_summary_flag:
             # make summaries of groups of docs
-            intermediate_summaries = [
-                self.summarization_chain.run(intermediate_summaries[i : i + group_size])
-                for i in range(0, len(intermediate_summaries), group_size)
-            ]
-            intermediate_summaries = [
-                Document(page_content=summary, metadata={"is_summary":True})
-                for summary in intermediate_summaries
-            ]
+            new_summaries = []
+            for i in range(0, len(intermediate_summaries), group_size):
+                group = intermediate_summaries[i : i + group_size]
+                group = list(map(lambda d: d.page_content, group))
+
+                summary = self.summarization_chain.run(
+                    separator + separator.join(group)
+                )
+                summary = Document(page_content=summary)
+                new_summaries.append(summary)
 
             # update list of all summaries
-            all_summaries = intermediate_summaries + all_summaries
+            all_summaries = new_summaries.copy() + all_summaries
+            intermediate_summaries = new_summaries
 
             # did we reach root summary?
             root_summary_flag = len(intermediate_summaries) == 1
@@ -191,6 +145,8 @@ class CheshireCat:
                 log(
                     f"Building summaries over {len(intermediate_summaries)} chunks. Please wait."
                 )
+
+        log(all_summaries)
 
         # return root summary and all intermediate summaries
         return all_summaries[0], all_summaries[1:]
@@ -248,32 +204,33 @@ class CheshireCat:
         docs = [summary] + intermediate_summaries + docs
         RabbitHole.store_documents(ccat=self, docs=docs, source=url)
 
+    def recall_relevant_memories_to_working_memory(self, user_message):
+        # TODO: should we compress convo history for better retrieval?
+
+        # HyDE
+        hyde_text, hyde_embedding = self.get_hyde_text_and_embedding(user_message)
+
+        # recall relevant memories (episodic)
+        episodic_memories = self.memory.vectors.episodic.recall_memories_from_embedding(
+            embedding=hyde_embedding
+        )
+        self.working_memory["episodic_memories"] = episodic_memories
+
+        # recall relevant memories (declarative)
+        declarative_memories = (
+            self.memory.vectors.declarative.recall_memories_from_embedding(
+                embedding=hyde_embedding
+            )
+        )
+        self.working_memory["declarative_memories"] = declarative_memories
+
     def __call__(self, user_message):
         if self.verbose:
             log(user_message)
 
-        hyde_text, hyde_embedding = self.get_hyde_text_and_embedding(user_message)
-
+        # recall episodic and declarative memories from vector collections and store them in working_memory
         try:
-            # recall relevant memories (episodic)
-            episodic_memory_content = (
-                self.memory.vectors.episodic.recall_memories_from_embedding(
-                    embedding=hyde_embedding
-                )
-            )
-            episodic_memory_formatted_content = self.format_memories_for_prompt(
-                episodic_memory_content
-            )
-
-            # recall relevant memories (declarative)
-            declarative_memory_content = (
-                self.memory.vectors.declarative.recall_memories_from_embedding(
-                    embedding=hyde_embedding
-                )
-            )
-            declarative_memory_formatted_content = self.format_memories_for_prompt(
-                declarative_memory_content
-            )
+            self.recall_relevant_memories_to_working_memory(user_message)
         except Exception as e:
             log(e)
             traceback.print_exc(e)
@@ -284,14 +241,33 @@ class CheshireCat:
                 "why": {},
             }
 
+        # format memories to be inserted in the prompt
+        episodic_memory_formatted_content = self.mad_hatter.execute_hook(
+            "format_episodic_memories_for_prompt",
+            self.working_memory["episodic_memories"],
+        )
+        declarative_memory_formatted_content = self.mad_hatter.execute_hook(
+            "format_declarative_memories_for_prompt",
+            self.working_memory["declarative_memories"],
+        )
+
+        # format conversation history to be inserted in the prompt
+        conversation_history_formatted_content = self.mad_hatter.execute_hook(
+            "format_conversation_history_for_prompt", self.working_memory["history"]
+        )
+
+        # load agent (will rebuild both agent and agent_executor based on context and plugins)
+        agent_executor = self.agent_manager.get_agent_executor()
+
         # reply with agent
         try:
-            cat_message = self.agent_executor(
+            cat_message = agent_executor(
                 {
                     "input": user_message,
                     "episodic_memory": episodic_memory_formatted_content,
                     "declarative_memory": declarative_memory_formatted_content,
-                    "chat_history": self.history,
+                    "chat_history": conversation_history_formatted_content,
+                    "ai_prefix": "AI",
                 }
             )
         except ValueError as e:
@@ -310,20 +286,18 @@ class CheshireCat:
             log(cat_message)
 
         # update conversation history
-        self.history += f"Human: {user_message}\n"
-        self.history += f'AI: {cat_message["output"]}\n'
+        self.working_memory.update_conversation_history(
+            who="Human", message=user_message
+        )
+        self.working_memory.update_conversation_history(
+            who="AI", message=cat_message["output"]
+        )
 
         # store user message in episodic memory
         # TODO: vectorize and store also conversation chunks (not raw dialog, but summarization)
         _ = self.memory.vectors.episodic.add_texts(
             [user_message],
-            [
-                {
-                    "source": "user",
-                    "when": time.time(),
-                    "text": user_message,
-                }
-            ],
+            [{"source": "user", "when": time.time()}],
         )
 
         # build data structure for output (response and why with memories)
@@ -331,16 +305,18 @@ class CheshireCat:
             "error": False,
             "content": cat_message["output"],
             "why": {
-                **cat_message,
+                "input": cat_message["input"],
+                "output": cat_message["output"],
+                "intermediate_steps": cat_message["intermediate_steps"],
                 "memory": {
                     "vectors": {
                         "episodic": [
                             dict(d[0]) | {"score": float(d[1])}
-                            for d in episodic_memory_content
+                            for d in self.working_memory["episodic_memories"]
                         ],
                         "declarative": [
                             dict(d[0]) | {"score": float(d[1])}
-                            for d in declarative_memory_content
+                            for d in self.working_memory["declarative_memories"]
                         ],
                     }
                 },