+++ core/cat/rabbit_hole.py
@@ -9,9 +9,9 @@ from langchain.text_splitter import RecursiveCharacterTextSplitter
 from starlette.datastructures import UploadFile
 from langchain.document_loaders import (
     PDFMinerLoader,
+    UnstructuredURLLoader,
     UnstructuredFileLoader,
     UnstructuredMarkdownLoader,
-    UnstructuredURLLoader,
 )
 from langchain.docstore.document import Document
 
@@ -19,7 +19,7 @@ from langchain.docstore.document import Document
 class RabbitHole:
     def __init__(self):
         pass
-    
+
     @staticmethod
     def url_to_docs(
         url: str,
@@ -63,7 +63,7 @@ class RabbitHole:
         """
 
         # Create temporary file
-        temp_file = tempfile.NamedTemporaryFile(dir=".", delete=False)
+        temp_file = tempfile.NamedTemporaryFile(dir="/tmp/", delete=False)
         temp_name = temp_file.name
 
         # Check type of incoming file.