+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
-## 0.0.2 ( 2023-04-22 )
+
+## 0.0.3 ( 2023-05-03 )
+
+### Enhancements
+
+* added more hooks to control prompting and allowed tools (see `core_plugin`)
+* added `cat.working_memory` to store temporary and arbitrary info (recent convo, recalled memories, data shared by plugins)
+* endpoints to erase memory contents (also completely wipe out memory)
+
+## 0.0.2 ( 2023-04-28 )
 
 ### Enhancements
 
@@ -8,6 +17,7 @@
 * plugin system `MadHatter` sorts hooks by priority
 * more hooks to control summarization prompt and reelaborate final reply to user
 * chat autofocus and easier home navigation in admin
+* support for Azure LLMs
 
 ## 0.0.1 ( 2023-04-20 )